+++ app/Http/Controllers/AdminMateriController.php
@@ -14,42 +14,58 @@ class AdminMateriController extends Controller
 
     public function index()
     {
-        $data = Materi::all();
-        return view('admin/materi',['dosen'=>$data]);
+        if(Session::get('username')) {
+            $data = Materi::all();
+            return view('admin/materi', ['dosen' => $data]);
+        }else{
+            return redirect('admin/auth');
+        }
     }
 
     public function show($id){
-        $data = Materi::find($id);
-        $dosen = Dosen::all();
-        return view('admin/edit_materi',['materi'=>$data, 'dosen' => $dosen]);
+        if(Session::get('username')) {
+            $data = Materi::find($id);
+            $dosen = Dosen::all();
+            return view('admin/edit_materi', ['materi' => $data, 'dosen' => $dosen]);
+        }else{
+            return redirect('admin/auth');
+        }
     }
 
     public function update(Request $request, $id)
     {
-        $this->validate($request, [
-            'topik' => 'required',
-            'isi_materi' => 'required',
-            'matkul' => 'required'
-        ]);
+        if(Session::get('username')) {
+            $this->validate($request, [
+                'topik' => 'required',
+                'isi_materi' => 'required',
+                'matkul' => 'required'
+            ]);
 
-        $data = Materi::find($id);
-        $data->topik = $request->input('topik');
-        $data->isi_materi = $request->input('isi_materi');
-        if($request->hasFile('file')){
-            $file = requet::file('file');
-            $filename = $file->getClientOriginalName();
-            $path = public_path().'/uploads/';
-            $data->video = $filename;
-            $file->move($path, $filename);
+            $data = Materi::find($id);
+            $data->topik = $request->input('topik');
+            $data->isi_materi = $request->input('isi_materi');
+            if ($request->hasFile('file')) {
+                $file = requet::file('file');
+                $filename = $file->getClientOriginalName();
+                $path = public_path() . '/uploads/';
+                $data->video = $filename;
+                $file->move($path, $filename);
+            }
+            $data->matkul = $request->input('matkul');
+            $data->save();
+            return redirect('/admin/materi');
+        }else{
+            return redirect('admin/auth');
         }
-        $data->matkul = $request->input('matkul');
-        $data->save();
-        return redirect('/admin/materi');
     }
 
     public function destroy($id){
-        $data = Materi::find($id);
-        $data->delete();
-        return redirect('/admin/materi');
+        if(Session::get('username')) {
+            $data = Materi::find($id);
+            $data->delete();
+            return redirect('/admin/materi');
+        }else{
+            return redirect('admin/auth');
+        }
     }
 }
